Formatting and minor details in landing page
The bullet points in installation did not render correctly, and the
doi link to Gorman et al paper was not formatted properly. Some
missing full stops are added.

Also, a somewhat nitpicky footnote is added to explain that although
Git-annex is written in Haskell we recommend using Python wheels for
convenience.