Prioritize CMD in the examples
The main difference is referring to USERPROFILE. The table with
command examples now lists both the CMD and PowerShell spelling.

IMO, using CMD gives the most "native" Windows experience, and avoids
having to change script execution policy for activating a virtual
environment (compared to PowerShell).

A note is added to say that CMD is recommended and we will assume CMD
in the examples -- that should not prevent anybody from using the
shell of their choice.