Add more windows-specific solutions
Echo command exists, but behaves differently on Windows: it would have
echoed both the "-e" flag and the double quotes. For multiline, (echo
this & echo that) was the best I could find in 5 minutes reading but
that does not have to be the only way.

In part 3 we shall now use %USERPROFILE% consistently.

For maximum correctness we could also adjust folder/file paths for
datalad drop, git annex info, etc. to be platform-specific across the
notebooks, but duplicating examples feels unnecessary -- both datalad
and git-annex (and Git, probably) accept UNIX paths in all these
situations, so the commands are perfectly copy-pastable.